requirements.txt
pyTelegramBotAPI
openai
python-dotenv